refactor: simplify main.rs
