+++ Cargo.toml
@@ -3,7 +3,7 @@ name = "rustywatch"
 authors = ["Adiatma Kamarudin <adiatma.mail@gmail.com>"]
 rust-version = "1.81.0"
 description = "Live reloading for any programing languages"
-version = "0.2.7"
+version = "0.2.8"
 edition = "2021"
 license = "MIT OR Apache-2.0"
 repository = "https://github.com/ak9024/rustywatch"
