+++ src/args.rs
@@ -1,21 +1,19 @@
 use clap::Parser;
 
 const TITLE: &str = r#"
-╭━━━╮╱╱╱╱╱╭╮╱╱╱╱╭╮╭╮╭╮╱╱╭╮╱╱╱╭╮
-┃╭━╮┃╱╱╱╱╭╯╰╮╱╱╱┃┃┃┃┃┃╱╭╯╰╮╱╱┃┃
-┃╰━╯┣╮╭┳━┻╮╭╋╮╱╭┫┃┃┃┃┣━┻╮╭╋━━┫╰━╮
-┃╭╮╭┫┃┃┃━━┫┃┃┃╱┃┃╰╯╰╯┃╭╮┃┃┃╭━┫╭╮┃
-┃┃┃╰┫╰╯┣━━┃╰┫╰━╯┣╮╭╮╭┫╭╮┃╰┫╰━┫┃┃┃
-╰╯╰━┻━━┻━━┻━┻━╮╭╯╰╯╰╯╰╯╰┻━┻━━┻╯╰╯
-╱╱╱╱╱╱╱╱╱╱╱╱╭━╯┃
-╱╱╱╱╱╱╱╱╱╱╱╱╰━━╯"#;
+ ____            _       __        __    _       _     
+|  _ \ _   _ ___| |_ _   \ \      / /_ _| |_ ___| |__  
+| |_) | | | / __| __| | | \ \ /\ / / _` | __/ __| '_ \ 
+|  _ <| |_| \__ \ |_| |_| |\ V  V / (_| | || (__| | | |
+|_| \_\\__,_|___/\__|\__, | \_/\_/ \__,_|\__\___|_| |_|
+                     |___/                             
+"#;
 
 pub fn title() {
     println!("{}", TITLE);
 }
 
 #[derive(Parser, Debug)]
-#[command(about = None)]
 #[clap(
     version,
     author = clap::crate_authors!("\n"),
